00-bootcamp-project/requirements.txt

google-cloud-bigquery==3.9.0
pandas==2.0.0
pyarrow==11.0.0